+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [next]
 
+- `FlyoutListTile` can be used outside of a flyout ([#650](https://github.com/bdlukaa/fluent_ui/issues/650))
 - Add uk localization ([#647](https://github.com/bdlukaa/fluent_ui/pull/647))
 - Ensure `fontFamily` is inherit in some widgets ([654](https://github.com/bdlukaa/fluent_ui/issues/654))
+- Add `Flyout.navigatorKey` ([#538](https://github.com/bdlukaa/fluent_ui/issues/538))
+- Add `Card.borderColor` ([#643](https://github.com/bdlukaa/fluent_ui/issues/643))
 
 ## 4.1.2